+++ packages/vital-test/src/shadow/@bodiless/vital-layout/Footer.ts
@@ -15,9 +15,12 @@
 import { vitalFooterBase, asFooterToken } from '@bodiless/vital-layout';
 import { addProps } from '@bodiless/fclasses';
 
+// Note: when used with the __vital__ package on vital-demo, that package takes precedence
+// over this one.  This test override has no effect and just an example.
+
 const Default = asFooterToken(vitalFooterBase.Default, {
   Behavior: {
-    Wrapper: addProps({ 'data-shadowed-by': '__vital__:Footer' }),
+    Wrapper: addProps({ 'data-shadowed-by': '__vitaltest__:Footer' }),
   },
 });